xox game developed with Kotlin
This project presents a classic XOX (Tic-Tac-Toe) game implemented in Kotlin, showcasing the versatility and simplicity of the language for game development. The game follows the traditional rules, providing an engaging and intuitive user experience. With an elegant and efficient Kotlin codebase, the project demonstrates the power of the language in crafting interactive applications. Whether you're a Kotlin enthusiast or a beginner, this XOX game project serves as an excellent example of creating a timeless and enjoyable gaming experience using modern programming techniques. Explore the source code on GitHub to delve into the intricacies of the implementation and learn more about Kotlin's capabilities in game development.